0.5.5

+++ app/src/main/java/com/example/find_work_it/common/autorization/model/Tokens.kt
@@ -2,7 +2,7 @@ package com.example.find_work_it.common.autorization.model
 
 data class Tokens(
     val access_token: String?,
-    val expires_in: Long?,
+    val expires_in: Int?,
     val refresh_token: String?,
     val token_type: String?
 )
